+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 #### 0.13.4
+- Add `download` link handling to the `FakeAssetService`.
 - [fix] Determining the resource id in `FakeAssetService` fails when the endpoint base path is not `/`.
 - [fix] `links` attribute in fake asset service response contains wrong URIs.